fix: refactored functions
