+++ lib/draw/BpmnRenderer.js
@@ -16,7 +16,10 @@ import {
 
 import {
   getLabel,
-  isLabel
+  isLabel,
+  getTextWidthForTextAnnotation,
+  changeBounds,
+  getTextAnnotationHeightFromElement,
 } from '../util/LabelUtil';
 
 import {
@@ -65,7 +68,6 @@ import {
 import { Ids } from 'ids';
 
 import { black } from './BpmnRenderUtil';
-import { measureTextLines } from '../features/modeling/BpmnUpdater';
 
 var markerIds = new Ids();
 
@@ -2212,7 +2214,7 @@ export default function BpmnRenderer(
         }
       });
 
-      height = getTextHeightForTextAnnotation(renderedLabel);
+      height = getTextAnnotationHeightFromElement(renderedLabel);
       textElement.setAttribute('height', height);
       width = getTextWidthForTextAnnotation(renderedLabel, width);
       element.width = width;
@@ -2444,46 +2446,4 @@ function pickAttrs(attrs, keys = []) {
 
     return pickedAttrs;
   }, {});
-}
-
-function changeBounds(newBounds, element) {
-  element.x = newBounds.x;
-  element.y = newBounds.y;
-  element.width = newBounds.width;
-  element.height = newBounds.height;
-}
-
-
-function getTextHeightForTextAnnotation(text) {
-  return Math.max(text.lastChild.y.animVal[0].value, 30) + 10;
-}
-
-function getTextWidthForTextAnnotation(text, width) {
-
-  if ((text.textContent === '')) {
-    return 100;
-  }
-
-  let lineLengths = Array.from(measureTextLines(text.textContent, getStyleFromTextannotation(text), width + 5, true))
-    .map(elem => elem.width);
-
-  return Math.round(Math.max(...lineLengths));
-}
-
-function getStyleFromTextannotation(text) {
-  const styleString = text.attributes.style.nodeValue;
-
-  let style = Object.fromEntries(
-    styleString
-      .split(';')
-      .map(part => part.trim())
-      .filter(Boolean)
-      .map(part => {
-        const [ key, value ] = part.split(':').map(s => s.trim());
-        return [ key, value ];
-      })
-  );
-  style['lineHeight'] = text.attributes.lineHeight.nodeValue;
-
-  return style;
 }
