+++ lib/features/modeling/BpmnUpdater.js
@@ -8,16 +8,22 @@ import { getBusinessObject, getDi, is } from '../../util/ModelUtil';
 
 import { isAny } from './util/ModelingUtil';
 
-import { getLabel, isLabel, isLabelExternal } from '../../util/LabelUtil';
-
-import { append as svgAppend, attr as svgAttr, create as svgCreate, remove as svgRemove } from 'tiny-svg';
+import {
+  getLabel,
+  getTextHeightForTextAnnotation,
+  isLabel,
+  isLabelExternal,
+  getConnectionGfx,
+  canHandleLabelResize,
+  getTextHeightForLabel,
+  getTextWidthForLabel
+} from '../../util/LabelUtil';
 
 import { isPlane } from '../../util/DrilldownUtil';
 
 import { delta } from 'diagram-js/lib/util/PositionUtil';
 
 import CommandInterceptor from 'diagram-js/lib/command/CommandInterceptor';
-import { assignStyle } from 'min-dom';
 
 /**
  * @typedef {import('diagram-js/lib/core/EventBus').default} EventBus
@@ -207,61 +213,6 @@ export default function BpmnUpdater(
     }
 
   });
-
-  function canHandleLabelResize(event) {
-    const element = event.element,
-
-          currentlabel = getLabel(element),
-
-          renderedLabel = getGfxLabel(event.gfx, currentlabel);
-
-    if (currentlabel !== renderedLabel) {
-      return true;
-    }
-    let renderedLabelHeight = getTextHeightForLabel(event),
-        renderedLabelWidth = getTextWidthForLabel(event);
-    if (hasMoreDiffThan(element, renderedLabelHeight, renderedLabelWidth, 5)) {
-      return true;
-    }
-
-    return false;
-  }
-
-  function getConnectionGfx(event, connection) {
-    const children = event.gfx.parentElement.parentElement.children;
-    for (let child of children) {
-      const dataElementId = child.firstChild.attributes?.getNamedItem('data-element-id').value;
-      if (connection.id === dataElementId) {
-        return child;
-      }
-    }
-  }
-
-  function getTextHeightForTextAnnotation(event) {
-    return Math.max(event.gfx.firstChild.children[1].lastChild.y.animVal[0].value, 30) + 10;
-  }
-
-  function getTextHeightForLabel(event) {
-    const text = event.gfx.firstChild.firstChild;
-    return Math.max(Math.round(text.lastChild.y.animVal[0].value), 15);
-  }
-
-  function getTextWidthForLabel(event) {
-    let lineLengths = Array.from(measureTextLines(getGfxLabel(event.gfx), getStyleFromLabel(event), event.element.width + 5))
-      .map(elem => elem.width);
-
-    return Math.round(Math.max(...lineLengths));
-  }
-
-  function getGfxLabel(gfx) {
-    return Array.from(gfx.firstChild.firstChild.children)
-      .map(child => child.textContent)
-      .join('');
-  }
-
-  function hasMoreDiffThan(element, height, width, padding) {
-    return Math.abs(element.width - width) > padding || Math.abs(element.height - height) > padding;
-  }
 
   // attach / detach connection
   function updateConnection(e) {
@@ -896,174 +847,4 @@ function getEmbeddedLabelBounds(shape) {
   }
 
   return label.get('bounds');
-}
-
-// ---------- helper svg --------------
-
-function getHelperSvg() {
-  let helperSvg = document.getElementById('helper-svg');
-
-  if (!helperSvg) {
-    helperSvg = svgCreate('svg');
-
-    svgAttr(helperSvg, { id: 'helper-svg' });
-
-    assignStyle(helperSvg, {
-      visibility: 'hidden',
-      position: 'fixed',
-      width: 0,
-      height: 0,
-    });
-
-    document.body.appendChild(helperSvg);
-  }
-
-  return helperSvg;
-}
-
-// ---------- text measurement ---------
-
-function getTextBBox(text, fakeText) {
-  fakeText.textContent = text;
-
-  try {
-    const emptyLine = text === '';
-    fakeText.textContent = emptyLine ? 'dummy' : text;
-
-    const raw = fakeText.getBBox();
-
-    const box = {
-      width: raw.width + raw.x * 2,
-      height: raw.height,
-    };
-
-    if (emptyLine) box.width = 0;
-
-    return box;
-  } catch (e) {
-    console.warn('BBox failed:', e);
-    return { width: 0, height: 0 };
-  }
-}
-
-// -------- shorten logic ----------
-
-const SOFT_BREAK = '\u00AD';
-
-function semanticShorten(line, maxLength) {
-  const parts = line.split(/(\s|-|\u00AD)/g);
-  let part;
-  const shortenedParts = [];
-  let length = 0;
-
-  while ((part = parts.shift())) {
-    if (part.length + length < maxLength) {
-      shortenedParts.push(part);
-      length += part.length;
-    } else {
-      if (part === '-' || part === SOFT_BREAK) {
-        shortenedParts.pop();
-      }
-      break;
-    }
-  }
-
-  const last = shortenedParts[shortenedParts.length - 1];
-  if (last === SOFT_BREAK) shortenedParts[shortenedParts.length - 1] = '-';
-
-  return shortenedParts.join('');
-}
-
-function shortenLine(line, width, maxWidth) {
-  const length = Math.max((line.length * maxWidth) / width, 1);
-  let shortened = semanticShorten(line, length);
-
-  if (!shortened) {
-    shortened = line.slice(0, Math.max(Math.round(length - 1), 1));
-  }
-  return shortened;
-}
-
-// ---------- layout line -------------
-
-function layoutNext(lines, maxWidth, fakeText) {
-  const originalLine = lines.shift();
-  let fitLine = originalLine;
-
-  for (;;) {
-    const bbox = getTextBBox(fitLine, fakeText);
-    bbox.width = fitLine ? bbox.width : 0;
-
-    // fits or too short to split
-    if (
-      fitLine === ' ' ||
-            fitLine === '' ||
-            bbox.width < Math.round(maxWidth) ||
-            fitLine.length < 2
-    ) {
-      return fit(lines, fitLine, originalLine, bbox);
-    }
-
-    // otherwise shorten and try again
-    fitLine = shortenLine(fitLine, bbox.width, maxWidth);
-  }
-}
-
-function fit(lines, fitLine, originalLine, bbox) {
-  if (fitLine.length < originalLine.length) {
-    const remainder = originalLine.slice(fitLine.length).trim();
-    lines.unshift(remainder);
-  }
-
-  // is doof
-  return {
-    text: fitLine,
-    width: Math.round(bbox.width),
-    height: bbox.height,
-  };
-}
-
-function getStyleFromLabel(event) {
-  const styleString = event.gfx.firstChild.firstChild.attributes.style.nodeValue;
-
-  let style = Object.fromEntries(
-    styleString
-      .split(';')
-      .map(part => part.trim())
-      .filter(Boolean)
-      .map(part => {
-        const [ key, value ] = part.split(':').map(s => s.trim());
-        return [ key, value ];
-      })
-  );
-  style['lineHeight'] = event.gfx.firstChild.firstChild.attributes.lineHeight.nodeValue;
-
-  return style;
-}
-
-// =====================================================
-//              PUBLIC: measureTextLines()
-// =====================================================
-
-export function measureTextLines(text, style, maxWidth, isTextAnnotation) {
-
-  // prepare helper text node
-  const helperText = svgCreate('text');
-  svgAttr(helperText, { x: 0, y: 0 });
-  svgAttr(helperText, style);
-
-  const helperSvg = getHelperSvg();
-  svgAppend(helperSvg, helperText);
-
-  const lines = text.split(/\u00AD?\r?\n/);
-  const results = [];
-
-  while (lines.length) {
-    const line = layoutNext(lines, maxWidth, helperText, isTextAnnotation);
-    results.push(line);
-  }
-
-  svgRemove(helperText);
-
-  return results;
 }
